+++ xrpl/asyncio/clients/json_rpc_base.py
@@ -2,11 +2,14 @@
 from __future__ import annotations
 
 from httpx import AsyncClient
+from typing_extensions import Final
 
 from xrpl.asyncio.clients.client import Client
 from xrpl.asyncio.clients.utils import json_to_response, request_to_json_rpc
 from xrpl.models.requests.request import Request
 from xrpl.models.response import Response
+
+_TIMEOUT: Final[float] = 10.0
 
 
 class JsonRpcBase(Client):
@@ -16,19 +19,9 @@ class JsonRpcBase(Client):
     :meta private:
     """
 
-    def __init__(self: JsonRpcBase, url: str) -> None:
-        """
-        Initialize this JsonRpcClient.
-
-        Arguments:
-            url: The url to which to connect.
-        """
-        self.url = url
-
     async def request_impl(self: JsonRpcBase, request: Request) -> Response:
         """
-        Asynchronously submit the request represented by the request_object to the
-        rippled node specified by this client's URL.
+        Base ``request_impl`` implementation for JSON RPC.
 
         Arguments:
             request: An object representing information about a rippled request.
@@ -38,7 +31,7 @@ class JsonRpcBase(Client):
 
         :meta private:
         """
-        async with AsyncClient(timeout=10.0) as http_client:
+        async with AsyncClient(timeout=_TIMEOUT) as http_client:
             response = await http_client.post(
                 self.url,
                 json=request_to_json_rpc(request),